Revert "Revert "refactor(suite-sync): remove redundant action definitions""
This reverts commit f0e67a077487e54372c179f0a2b7908a6e42c4fb.
