+++ suite-native/module-dev-utils/src/components/SuiteSyncRelaySettings.tsx
@@ -4,7 +4,8 @@ import {
     DEFAULT_SUITE_SYNC_RELAY_URL,
     selectIsSuiteSyncDebugEnabled,
     selectSuiteSyncRelayUrl,
-    suiteSyncActions,
+    updateIsFeatureSuiteSyncAvailable,
+    updateSuiteSyncDebugEnabled,
 } from '@suite-common/suite-sync';
 import { yup } from '@suite-common/validators';
 import { Button, Card, CheckBox, HStack, Text, VStack } from '@suite-native/atoms';
@@ -38,9 +39,17 @@ export const SuiteSyncRelaySettings = () => {
         });
     });
 
+    const handleSuiteSyncEnableToggle = () => {
+        dispatch(
+            updateIsFeatureSuiteSyncAvailable({
+                isShownInSettings: !isFeatureSuiteSyncEnabled,
+            }),
+        );
+    };
+
     const handleSuiteSyncDebugToggle = () => {
         dispatch(
-            suiteSyncActions.updateSuiteSyncDebugEnabled({
+            updateSuiteSyncDebugEnabled({
                 isEnabled: !isSuiteSyncDebugEnabled,
             }),
         );
